requirements.txt
requests>=2.28.0
lxml>=4.9.0
PyYAML>=6.0
tqdm>=4.64.0
ipinfo>=4.4.0